Adjust supported Ruby versions
Ruby 2.4 is now end-of-life. As a result of this, the `google-protobuf`
gem is no longer releasing artifacts compatible with this version.
(https://github.com/protocolbuffers/protobuf/pull/7453)

In CI we encounter this error from `bundle install`:
```
google-protobuf-3.12.0-x86_64-linux requires ruby version >= 2.5, which is
incompatible with the current version, ruby 2.4.6p354
```

Therefore remove Ruby 2.4.6 as a tested version in Travis CI.

Additionally remove the constraint on patch versions in the Travis
config, so that we'll use the latest patch version available for each
release branch.